requirements.txt
Flask>=2.2.5
Flask-Migrate>=3.1.0
Flask-SQLAlchemy>=3.0.3
python-dotenv>=1.0.0
APScheduler>=3.10.5
beautifulsoup4>=4.12.2
pytz>=2023.3
selenium>=4.10.0
webdriver-manager>=3.8.6
openai>=1.3.0
requests>=2.31.0
psycopg2-binary>=2.9.7
gunicorn>=20.1.0
Pillow>=10.0.0
lxml>=4.9.3
